chore: Add split useGa and useGlean into separate files
